move concatdataloader out
