+++ scvi/dataloaders/__init__.py
@@ -1,3 +1,5 @@
-from ._scvi_dataloader import ConcatDataLoader, ScviDataLoader
+from ._concat_dataloader import ConcatDataLoader
+from ._scvi_dataloader import ScviDataLoader
+
 
 __all__ = ["ScviDataLoader", "ConcatDataLoader"]
